[refactor] Move serilalizers files in app
공통으로 관리되던 시리얼라이져들을 각 앱에 배치했습니다.
각 앱별로만 사용하는 시리얼라이져가 분리된 구조가 더 직관적이라 판단돼 진행했습니
