--- blog/post/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from .models import Post, Comment
-
-admin.site.register(Post)
-admin.site.register(Comment)
-# Register your models here.
