+++ blog/post/serializers.py
@@ -1,12 +1,12 @@
 from rest_framework import serializers
-from post.models import Post
+from .models import Post
 from user.models import User
 
 
 class PostSerializer(serializers.ModelSerializer):
     class Meta:
         model = Post
-        fields = ["id", "title", "content", "author"]
+        fields = ["id", "title", "content", "author", "created_at"]
 
     id = serializers.IntegerField(read_only=True)
     title = serializers.CharField(max_length=50, required=True)
